flask_app/requirements.txt
annotated-types==0.7.0
asttokens==2.4.1
attrs==25.3.0
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
colorlog==6.8.2
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
dnspython==2.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
executing==2.0.1
Faker==26.3.0
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.53.1
git-filter-repo==2.45.0
gitdb==4.0.11
GitPython==3.1.43
graphviz==0.20.3
gunicorn==23.0.0
icecream==2.1.3
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
jellyfish==1.1.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
load-dotenv==0.1.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
mdurl==0.1.2
murmurhash==1.0.10
networkx==3.3
nltk==3.8.1
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.4.0
preshed==3.0.9
psutil==6.0.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pymongo==4.10.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.36.2
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rpds-py==0.23.1
scikit-learn==1.5.1
scipy==1.14.0
segtok==1.5.11
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
tabulate==0.9.0
thinc==8.2.5
threadpoolctl==3.5.0
tqdm==4.66.5
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.3
wrapt==1.16.0
yake==0.4.8
youtube-transcript-api==0.6.2
zipp==3.20.0
